Several improvements to the rheology script, added a new script to auto generate tests, WIP
1. Rheology:
Added a help file to the script.
Added test files
Added a new automatic fitting method, which divides the error by the number of points to select a larger range of data.
Added global variables to the script as general settings.
2. Automatic test maker
Still a very early WIP. It is supposed to generate several files with randomized parameters which specific answers. Doesn't work at all at the moment.